Improve library ergonomics
When testing out the bar there were a few ergonomic issues. These have
been fixed.

These adjustments have been made:
 * New `redraw` method that allows not redrawing a component at all
 * Add default implementations for the Component Trait's methods
 * Fix small bug where `comp_index` was incorrect because of filtering
 * Add yoffset to text again

+++ src/bar.rs
@@ -27,7 +27,7 @@ pub struct Bar {
     pub(crate) format24: u32,
     pub(crate) color: Color,
     pub(crate) component_ids: [u32; 3],
-    pub(crate) text_yoffset: f64,
+    pub(crate) text_yoffset: i16,
 }
 
 impl Bar {
